code/produce/requirements.txt
# requirements.txt
kafka-python
SPARQLWrapper
docker-host-ip
names
